Extract internal item classes and tree model to individual files and convert comments to javadoc
